+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to the "vz-like-keymap" extension will be documented in this
 ## [Unreleased]
 ### Added
 - Cursor keys with no modifier keys for cursor movement without leaving selection mode.
+### Fixed
+- Not leaving selection mode after Cut action (Ctrl+Y).
 
 ## [0.1.0] - 2019-10-24
 ### Added